PR-16 Created MyCategoryTypeExposedDropdownMenuBox
