+++ app/src/main/java/com/pbs/expenseApp/ui/components/AppScaffold.kt
@@ -5,6 +5,7 @@ import androidx.compose.runtime.Composable
 import com.pbs.expenseApp.appstate.rememberAppState
 import com.pbs.expenseApp.navigation.AppNavigationGraph
 import com.pbs.expenseApp.ui.composables.MyAddCategoryFab
+import com.pbs.expenseApp.ui.composables.MyAppNavigationBar
 
 @Composable
 fun AppScaffold() {
@@ -17,7 +18,7 @@ fun AppScaffold() {
         },
         bottomBar = {
             if (appState.shouldShowBottomBar) {
-                AppNavigationBar(appState = appState)
+                MyAppNavigationBar(appState = appState)
             }
         }
     ) { innerPadding ->
